+++ src/Factory/InputFactory.php
@@ -4,10 +4,9 @@ declare(strict_types=1);
 
 namespace Sfmok\RequestInput\Factory;
 
-use Sfmok\RequestInput\Attribute\Input;
 use Sfmok\RequestInput\Exception\DeserializationException;
-use Sfmok\RequestInput\InputInterface;
 use Sfmok\RequestInput\Exception\ValidationException;
+use Sfmok\RequestInput\InputInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Exception\UnsupportedMediaTypeHttpException;
 use Symfony\Component\Serializer\Exception\UnexpectedValueException;
@@ -24,15 +23,10 @@ final class InputFactory implements InputFactoryInterface
         private ValidatorInterface $validator,
         private bool $skipValidation,
         private array $inputFormats
-    ) {
-    }
+    ) {}
 
     public function createFromRequest(Request $request, string $type): iterable
     {
-        if (!is_subclass_of($type, InputInterface::class)) {
-            return [];
-        }
-
         if (!is_subclass_of($type, InputInterface::class)) {
             return [];
         }
@@ -52,18 +46,12 @@ final class InputFactory implements InputFactoryInterface
 
         $data = $request->getContent();
         $format = $request->getContentTypeFormat();
-        if (Input::INPUT_FORM_FORMAT === $format) {
-            @trigger_error("The format 'form' is deprecated and will be removed in version 2.0. Use 'symfony/form' component instead.", \E_USER_DEPRECATED);
-            $data = json_encode($request->request->all());
-            $format = Input::INPUT_JSON_FORMAT;
-        }
 
         try {
             $input = $this->serializer->deserialize($data, $type, $format, $inputMetadata?->getContext() ?? []);
         } catch (UnexpectedValueException $exception) {
             throw new DeserializationException('Deserialization Failed', $exception);
         }
-
 
         if (!$this->skipValidation) {
             $violations = $this->validator->validate($input, null, $inputMetadata?->getGroups() ?? ['Default']);