+++ src/Factory/InputFactoryInterface.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Sfmok\RequestInput\Factory;
 
-use Sfmok\RequestInput\InputInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 interface InputFactoryInterface