+++ tests/Fixtures/Controller/TestController.php
@@ -9,21 +9,13 @@ use Sfmok\RequestInput\Attribute\Input;
 class TestController
 {
     #[Input(format: 'json', groups: ['bar'], context: ['groups' => ['bar']])]
-    public function __invoke(): void
-    {
-    }
+    public function __invoke(): void {}
 
     #[Input(format: 'json', groups: ['foo'], context: ['groups' => ['foo']])]
-    public function testWithInput(): void
-    {
-    }
+    public function testWithInput(): void {}
 
-    public function testWithoutInput(): void
-    {
-    }
+    public function testWithoutInput(): void {}
 
     #[Input(format: 'unsupported')]
-    public function testWithInputUnsupportedFormat(): void
-    {
-    }
+    public function testWithInputUnsupportedFormat(): void {}
 }